Wait for a condition in the sidecar work-loop tests, not a fixed delay
A test slept 200 ms and then asserted the claim had been made, the job
executed and the loop had asked again. On a loaded macOS CI runner none
of that had happened yet, so a green suite reported five failures for a
defect that did not exist.

waitFor moves out of SidecarSessionTests into its own file and both
fragile work-loop tests use it. The drained-worker test still needs a
fixed pause to prove a claim was not made, but now waits for the loop to
be demonstrably running first rather than possibly looking too early.
